+++ packages/std/src/query/bank.rs
@@ -3,6 +3,8 @@ use serde::{Deserialize, Serialize};
 
 use crate::Coin;
 
+use super::query_response::QueryResponseType;
+
 #[non_exhaustive]
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
@@ -22,7 +24,7 @@ pub enum BankQuery {
 }
 
 #[cfg(feature = "cosmwasm_1_1")]
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+#[derive(Serialize, Deserialize, Clone, Debug, Default, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 #[non_exhaustive]
 pub struct SupplyResponse {
@@ -32,18 +34,9 @@ pub struct SupplyResponse {
 }
 
 #[cfg(feature = "cosmwasm_1_1")]
-impl SupplyResponse {
-    /// Constructor for testing frameworks such as cw-multi-test.
-    /// This is required because query response types should be #[non_exhaustive].
-    /// As a contract developer you should not need this constructor since
-    /// query responses are constructed for you via deserialization.
-    #[doc(hidden)]
-    pub fn new(amount: Coin) -> Self {
-        Self { amount }
-    }
-}
+impl QueryResponseType for SupplyResponse {}
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+#[derive(Serialize, Deserialize, Clone, Debug, Default, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub struct BalanceResponse {
     /// Always returns a Coin with the requested denom.
@@ -51,9 +44,13 @@ pub struct BalanceResponse {
     pub amount: Coin,
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+impl QueryResponseType for BalanceResponse {}
+
+#[derive(Serialize, Deserialize, Clone, Debug, Default, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub struct AllBalanceResponse {
     /// Returns all non-zero coins held by this account.
     pub amount: Vec<Coin>,
 }
+
+impl QueryResponseType for AllBalanceResponse {}