+++ CHANGELOG.md
@@ -23,7 +23,8 @@ and this project adheres to
 - cosmwasm-std: Upgrade `serde-json-wasm` dependency to 0.5.0 which adds map
   support to `to_vec`/`to_binary` and friends.
 - cosmwasm-std: Implement `AsRef<[u8]>` for `Binary` and `HexBinary` ([#1550]).
-- cosmwasm-std: Add constructor `SupplyResponse::new` ([#1552]).
+- cosmwasm-std: Allow constructing `SupplyResponse` via a `Default`
+  implementation ([#1552], [#1560]).
 
 [#1436]: https://github.com/CosmWasm/cosmwasm/issues/1436
 [#1437]: https://github.com/CosmWasm/cosmwasm/issues/1437
@@ -33,6 +34,7 @@ and this project adheres to
 [#1550]: https://github.com/CosmWasm/cosmwasm/issues/1550
 [#1552]: https://github.com/CosmWasm/cosmwasm/pull/1552
 [#1554]: https://github.com/CosmWasm/cosmwasm/pull/1554
+[#1560]: https://github.com/CosmWasm/cosmwasm/pull/1560
 
 ### Changed